+++ src/uncurried/match.js
@@ -1,0 +1,3 @@
+export default function match (r, s) {
+  return r.test(s)
+}